+++ lib/features/mypage/data/util/get_mento_level.dart
@@ -1,0 +1,9 @@
+import 'package:keodam/features/mypage/data/model/mento_level.dart';
+import 'package:keodam/features/mypage/data/model/mento_level_table.dart';
+
+MentoLevelInfo getMentoLevelInfo(int exp) {
+  return levelTable.firstWhere(
+    (level) => level.containsExp(exp),
+    orElse: () => levelTable.last,
+  );
+}
